fix go type errors in 1800

+++ src/standalone_processes/1800__dots-to-papers.go
@@ -558,10 +558,10 @@ func getGetPaperIdFromColors3(colors [][3]int, dotCodes8400 []string) (int, int,
 	// 	}
 	// }
 	// Assign 3 remaining colors
-	darkLuminance := 0.2126*colors[idealColorsToDotIndex[3]][0] + 0.7152*colors[idealColorsToDotIndex[3]][1] + 0.0722*colors[idealColorsToDotIndex[3]][2]
+	darkLuminance := 0.2126*float64(colors[idealColorsToDotIndex[3]][0]) + 0.7152*float64(colors[idealColorsToDotIndex[3]][1]) + 0.0722*float64(colors[idealColorsToDotIndex[3]][2])
 	minNonDarkLuminance := 99999.0
 	for i := 0; i < 3; i++ {
-		nonDarkLuminance := 0.2126*colors[idealColorsToDotIndex[i]][0] + 0.7152*colors[idealColorsToDotIndex[i]][1] + 0.0722*colors[idealColorsToDotIndex[i]][2]
+		nonDarkLuminance := 0.2126*float64(colors[idealColorsToDotIndex[i]][0]) + 0.7152*float64(colors[idealColorsToDotIndex[i]][1]) + 0.0722*float64(colors[idealColorsToDotIndex[i]][2])
 		if nonDarkLuminance < minNonDarkLuminance {
 			minNonDarkLuminance = nonDarkLuminance
 		}
@@ -571,7 +571,7 @@ func getGetPaperIdFromColors3(colors [][3]int, dotCodes8400 []string) (int, int,
 
 	for i, colorData := range colors {
 		if i != idealColorsToDotIndex[0] && i != idealColorsToDotIndex[1] && i != idealColorsToDotIndex[2] && i != idealColorsToDotIndex[3] {
-			luminance := 0.2126*colorData[0] + 0.7152*colorData[1] + 0.0722*colorData[2]
+			luminance := 0.2126*float64(colorData[0]) + 0.7152*float64(colorData[1]) + 0.0722*float64(colorData[2])
 			if luminance < luminanceThreshold {
 				matchedColors[i] = 3
 			} else {
